server/requirements.txt
pillow
fastapi[standard]
tensorflow
numpy
